Implement wayland Core.graceful_shutdown
And move the existing implementation (which is x11-specific) from
manager.py to the x11 Core class. A non-abstract method is added to the
base core.
