+++ libqtile/core/manager.py
@@ -28,7 +28,6 @@ import shutil
 import signal
 import subprocess
 import tempfile
-import time
 from typing import Dict, List, Optional, Tuple, Union
 
 import xcffib
@@ -234,7 +233,7 @@ class Qtile(CommandObject):
     def stop(self):
         hook.fire("shutdown")
         lifecycle.behavior = lifecycle.behavior.TERMINATE
-        self.graceful_shutdown()
+        self.core.graceful_shutdown()
         self._stop()
 
     def restart(self):
@@ -563,50 +562,6 @@ class Qtile(CommandObject):
                 c.group.remove(c)
             del self.windows_map[win]
             self.core.update_client_list(self.windows_map)
-
-    def graceful_shutdown(self):
-        """
-        Try and gracefully shutdown windows before exiting with SIGTERM, vs.
-        just closing the X session and having the X server send them all
-        SIGKILL.
-        """
-
-        def get_interesting_pid(win):
-            # We don't need to kill Internal or Static windows, they're qtile
-            # managed and don't have any state.
-            if not isinstance(win, base.Window):
-                return None
-            try:
-                return win.window.get_net_wm_pid()
-            except Exception:
-                logger.exception("Got an exception in getting the window pid")
-                return None
-        pids = map(get_interesting_pid, self.windows_map.values())
-        pids = list(filter(lambda x: x is not None, pids))
-
-        # Give the windows a chance to shut down nicely.
-        for pid in pids:
-            try:
-                os.kill(pid, signal.SIGTERM)
-            except OSError:
-                # might have died recently
-                pass
-
-        def still_alive(pid):
-            # most pids will not be children, so we can't use wait()
-            try:
-                os.kill(pid, 0)
-                return True
-            except OSError:
-                return False
-
-        # give everyone a little time to exit and write their state. but don't
-        # sleep forever (1s).
-        for i in range(10):
-            pids = list(filter(still_alive, pids))
-            if len(pids) == 0:
-                break
-            time.sleep(0.1)
 
     def find_screen(self, x, y):
         """Find a screen based on the x and y offset"""
